consolidated utils slightly
