--- ambersim/utils/_internal_utils.py
@@ -1,19 +1,0 @@
-from pathlib import Path
-from typing import Union
-
-from ambersim import ROOT
-
-
-def _check_filepath(filepath: Union[str, Path]) -> str:
-    """Checks validity of a filepath for model loading."""
-    assert isinstance(filepath, (str, Path))
-
-    # checking whether file exists
-    if isinstance(filepath, str):
-        filepath = Path(filepath)  # global/local
-    if not filepath.exists():
-        filepath = ROOT / filepath  # repo root
-        if not filepath.exists():
-            raise ValueError("The model file doesn't exist at the specified path!")
-    filepath = str(filepath)
-    return filepath
