+++ ambersim/utils/io_utils.py
@@ -7,7 +7,23 @@ import trimesh
 from dm_control import mjcf
 from mujoco import mjx
 
-from ambersim.utils._internal_utils import _check_filepath
+from ambersim import ROOT
+
+
+def _check_filepath(filepath: Union[str, Path]) -> str:
+    """Checks validity of a filepath for model loading."""
+    assert isinstance(filepath, (str, Path))
+
+    # checking whether file exists
+    if isinstance(filepath, str):
+        filepath = Path(filepath)  # global/local
+    if not filepath.exists():
+        filepath = ROOT / filepath  # repo root
+        if not filepath.exists():
+            raise ValueError("The model file doesn't exist at the specified path!")
+    filepath = str(filepath)
+    return filepath
+
 
 # ############# #
 # MODEL LOADING #
